Introduce RegWord type
On majority of architectures size of long matches register width.
On mips n32 size of long is 32 bits and register width is 64 bits.
Valgrind is written with assumption that long size matches register
width. This is the reason why both UWord for Valgrind and HWord for VEX
match size of long. Long size differs from register size on mips n32 ABI.

Introducing RegWord type that will match size of registers.

Part of the changes required for BZ issue - #345763.

Contributed by:
  Tamara Vlahovic and Dimitrije Nikolic.

+++ coregrind/m_libcsignal.c
@@ -92,7 +92,7 @@ Bool VG_(isemptysigset)( const vki_sigset_t* set )
       }
       else {
          /* Partial word check. */
-         ULong mask = (1UL << (_VKI_MAXSIG % _VKI_NSIG_BPW)) - 1;
+         ULong mask = ((ULong)1UL << (_VKI_MAXSIG % _VKI_NSIG_BPW)) - 1;
          if ((set->sig[i] & mask) != 0) return False;
          break;
       }
@@ -111,7 +111,7 @@ Bool VG_(isfullsigset)( const vki_sigset_t* set )
       }
       else {
          /* Partial word check. */
-         ULong mask = (1UL << (_VKI_MAXSIG % _VKI_NSIG_BPW)) - 1;
+         ULong mask = ((ULong)1UL << (_VKI_MAXSIG % _VKI_NSIG_BPW)) - 1;
          if ((set->sig[i] & mask) != mask) return False;
          break;
       }
@@ -130,7 +130,7 @@ Bool VG_(iseqsigset)( const vki_sigset_t* set1, const vki_sigset_t* set2 )
       }
       else {
          /* Partial word comparison. */
-         ULong mask = (1UL << (_VKI_MAXSIG % _VKI_NSIG_BPW)) - 1;
+         ULong mask = ((ULong)1UL << (_VKI_MAXSIG % _VKI_NSIG_BPW)) - 1;
          if ((set1->sig[i] & mask) != (set2->sig[i] & mask)) return False;
          break;
       }
